Refactor: 주문 스냅샷 저장으로 수정 결제 전에는 상품으로부터 실시간으로 정보를 가져옴 쿠폰 만료 전략 수정 쿠폰 만료를 상태로 저장하지 않고 조회시 만료일자를 통해 걸러내도록 변경

+++ src/main/java/com/example/pillyohae/coupon/repository/IssuedCouponQueryRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface IssuedCouponQueryRepository {
     // 유저가 가진 쿠폰중에 최소사용금액이 결제 총액보다 낮은경우를 찾음
-    List<FindCouponListToUseResponseDto.CouponInfo> findCouponListToUse(Double price, Long userId);
+    List<FindCouponListToUseResponseDto.CouponInfo> findCouponListToUse(Long price, Long userId);
 }
